+++ pytgbot/api_types/sendable/__init__.py
@@ -6,7 +6,11 @@ from luckydonaldUtils.logger import logging
 from pytgbot.api_types import TgBotApiObject
 
 __author__ = 'luckydonald'
-__all__ = ["inline", "reply_markup", "Sendable", "InputFile", "InputFileFromURL", "InputFileFromDisk"]
+__all__ = ["inline", "reply_markup", "files", "Sendable", "InputFile", "InputFileFromURL", "InputFileFromDisk"]
+
+# UPCOMING CHANGE IN v2.2.0:
+# from .files import InputFile, InputFileFromURL, InputFileFromDisk  # backwards compatibility, before v2.1.5
+
 logger = logging.getLogger(__name__)
 
 
@@ -18,14 +22,37 @@ class Sendable(TgBotApiObject):
 
 
 class InputFile(object):
-    def __init__(self, file_blob, file_name=None, file_mime=None):
-        super(InputFile, self).__init__()
+    def __init__(self, file_blob, file_name="file.unknown", file_mime=None):
+        super(InputFile, self).__init__()#
+        if not file_blob:
+            raise ValueError("The file content (file_blob argument) is required to be non-empty.")
+        # end if
+        if not file_name:
+            raise ValueError("Cannot have empty name (file_name argument).")
+        # end
+
         self.file_blob = file_blob
-        self.file_name = file_name if file_name else None
-        self.file_mime = file_mime if file_mime else None
-    # end def __init__
+        self.file_name = file_name
+        if file_mime:
+            self.file_mime = file_mime
+        else:
+            self.update_mime_from_blob()
+        # end if
+    # end def
+
+    def update_mime_from_blob(self):
+        """
+        Calculates the mime type from self.file_blob
+        :return:
+        """
+        if not self.file_mime:
+            import magic  # pip install python-magic
+            self.file_mime = magic.from_buffer(self.file_blob, mime=True)
+        # end if
+    # end def
 
     def get_request_files(self, var_name):
+        self.update_mime_from_blob(self)
         return {var_name: (self.file_name, self.file_blob, self.file_mime)}
     # end def get_request_files
 # end class InputFile
@@ -36,7 +63,11 @@ class InputFileFromDisk(InputFile):
         super(InputFile, self).__init__()
         self.file_path = file_path
         self.file_name = file_name if file_name else path.basename(file_path)
-        self.file_mime = file_mime if file_mime else None
+        if file_mime:
+            self.file_mime = file_mime
+        else:
+            self.update_mime_from_blob(self)
+        # end if file_mime
     # end def __init__
 
     def get_request_files(self, var_name):
@@ -48,15 +79,29 @@ class InputFileFromDisk(InputFile):
 class InputFileFromURL(InputFile):
     def __init__(self, file_url, file_name=None, file_mime=None):
         super(InputFileFromURL, self).__init__(None, file_name, file_mime)
+        # URL
         self.file_url = file_url
-        self.file = requests.get(file_url)
-        self.file_name = file_name if file_name else None
-        self.file_mime = file_mime if file_mime else None
-        self.url_to_name_and_mime()
+
+        # BLOB
+        self.update_blob_from_url()
+
+        # NAME
+        if file_name:
+            self.file_name = file_name
+        else:
+            self.update_name_from_url()
+        # end if
+
+        # MIME
+        if file_mime:
+            self.file_mime = file_name
+        else:
+            self.update_mime_from_blob()
+        # end if
     # end def __init__
 
     def __str__(self):
-        file_size = len(self.file.content)
+        file_size = len(self.file_blob)
         return (
             "{clazz_name!s}("
              "file_url={self.file_url!r}, file_name={self.file_name!r}, file_mime={self.file_mime!r}, "
@@ -64,7 +109,20 @@ class InputFileFromURL(InputFile):
         ).format(clazz_name=self.__class__.__name__, file_size=file_size, self=self)
     # end def __str__
 
-    def url_to_name_and_mime(self):
+    def update_blob_from_url(self):
+        """
+        Loads the file from the url to self.file and self.file_blob
+        :return:
+        """
+        self.file = requests.get(self.file_url)
+        self.file_blob = self.file.content
+    # end def
+
+    def update_name_from_url(self):
+        """
+        Cuts a name from an url
+        :return:
+        """
         if not self.file_name:
             import os  # http://stackoverflow.com/a/18727481/3423324
             try:  # python 2:
@@ -75,15 +133,11 @@ class InputFileFromURL(InputFile):
             path_part = urlparse(self.file_url).path
             self.file_name = os.path.basename(path_part)
         # end if
-        if not self.file_mime:
-            import magic  # pip install python-magic
-            self.file_mime = magic.from_buffer(requests.get(self.file_url).content, mime=True)
-        # end if
     # end def
 
 
 
     def get_request_files(self, var_name):
-        return {var_name: (self.file_name, self.file.content, self.file_mime)}
+        return {var_name: (self.file_name, self.file_blob, self.file_mime)}
     # end def get_request_files
 # end class InputFileFromURL