+++ pytgbot/api_types/sendable/files.py
@@ -1,0 +1,5 @@
+# -*- coding: utf-8 -*-
+__author__ = 'luckydonald'
+__all__ = ["InputFile", "InputFileFromURL", "InputFileFromDisk"]
+
+from . import InputFile, InputFileFromDisk, InputFileFromURL  # will be moved here in 2.2.0